move some tracing related assertions to separate module for tenant and timeline
